requirements.txt
Flask
torch
torchvision
timm
pillow
numpy
opencv-python
pytorch-grad-cam
